backend/requirements.txt
fastapi
uvicorn
requests
openai
python-multipart
python-dotenv
